+++ xlators/performance/io-cache/src/io-cache.h
@@ -131,7 +131,7 @@ struct ioc_inode {
 
 struct ioc_table {
   size_t page_size;
-  uint64_t cache_size;
+  size_t cache_size;
   uint64_t cache_used;
   struct list_head inodes; /* list of inodes cached */
   struct list_head active;
